Restructure files. Replace '.' imports with '@'.
