+++ README.md
@@ -1,16 +1,17 @@
 # CaveBoy
+
 A graphical editor for CoilSnake projects.
 
 ## Progress
 
-This project is still under development and does not yet have a functional release. 
+This project is still under development and does not yet have a functional release.
 
-- [ ] Milestone 1: Decode and encode *.fts files.
+- [ ] Milestone: Decode and encode \*.fts files.
   - [x] Colors
   - [ ] Minitile Palettes ("Subpalettes")
   - [ ] Minitile Layers
   - [ ] Minitiles
-  - [ ] Arrangment Cells 
+  - [ ] Arrangment Cells
   - [ ] Arrangements ("Tiles"/"Metatiles"/"Blocks")
   - [ ] Tilesets
   - [ ] Palette Sets ("Palettes"/"Palette Groups")
