+++ src/script/data/coilsnake-literal/CoilSnakeColorString.ts
