+++ tests/unit/script/data/game-object/Color.spec.ts
